+++ src/com/intellij/idea/plugin/hybris/project/view/HybrisProjectView.kt
@@ -25,11 +25,14 @@ import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.common.yExtensionName
 import com.intellij.idea.plugin.hybris.facet.YFacet
+import com.intellij.idea.plugin.hybris.project.services.HybrisProjectService
 import com.intellij.idea.plugin.hybris.settings.HybrisApplicationSettingsComponent
 import com.intellij.idea.plugin.hybris.settings.HybrisProjectSettingsComponent
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectRootManager
+import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.ui.SimpleTextAttributes
 import java.io.File
@@ -102,14 +105,22 @@ open class HybrisProjectView(val project: Project) : TreeStructureProvider, Dumb
         val otherNodes = mutableListOf<AbstractTreeNode<*>>()
         val treeNodes = mutableListOf<AbstractTreeNode<*>>()
 
+        val projectService = ApplicationManager.getApplication().getService(HybrisProjectService::class.java)
         val projectRootManager = ProjectRootManager.getInstance(project)
         for (child in children) {
             if (child is PsiDirectoryNode) {
-                val yFacet = child.virtualFile
-                    ?.let { projectRootManager.fileIndex.getModuleForFile(it) }
+                val virtualFile = child.virtualFile
+                    ?: continue
+
+                val file = VfsUtil.virtualToIoFile(virtualFile)
+                val yFacet = projectRootManager.fileIndex.getModuleForFile(virtualFile)
                     ?.let { YFacet.get(it) }
 
-                if (yFacet == null) {
+                if (yFacet == null && (projectService.isGradleModule(file)
+                        || projectService.isEclipseModule(file)
+                        || projectService.isGradleKtsModule(file)
+                        || projectService.isMavenModule(file))
+                ) {
                     otherNodes.add(child)
                 } else {
                     treeNodes.add(child)